src/requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
cryptography==2.1.4
et-xmlfile==1.0.1
googlemaps==2.5.1
idna==2.6
jdcal==1.3
openpyxl==2.5.0
protobuf==3.3.0
py3-ortools==6.4.4495
pycparser==2.18
pyOpenSSL==17.5.0
python-dateutil==2.7.0
requests==2.18.4
sip==4.19.8
six==1.11.0
urllib3==1.22
